fix: Use appropriate methods from ResultSet
Instead of calling getObject and inspecting the type.

This makes the parsers more consistent. And restores the previous behavior of ColReaders.

This makes it so longOption, bigDecimalOption, and stringOption, etc. may not throw exceptions in
cases where it did before. But I think this is probably the behavior that would be expected.