feat!: Fatal warnings
Fix all warnings and enable fatal warnings.

Also remove the dependency on scala-compat.

This is mostly backwards compatible, but won't work with versions of scala before 2.13.

It also adds an implicit conversion from SqlRow => A to RowParser[A], which allows continuing
to pass in a function as the argument instead of an implicit RowParser, but might be subtly
different in some cases.