Make classes more focused
There was some ambiguity in how some things should work (specifically,
`asScalarOption` and `as` in SqlResult) due to the very general nature
of things extending SqlResultWrapper. SqlResultWrapper made all the
methods on ResultSet available, which resulted in SqlRow and SqlResult
both being usable like a ResultSet. That's not their purpose though—
SqlRow represents _a single row_ and shouldn't expose implementation
details allowing it to access the whole result set (similarly, RowParser
should parse _a single row_ and never parse or be able to parse or
access anything beyond that row) and SqlResult represents the result set
as a whole, abstracting away the underlying implementation details.
These two classes _should not_ be interchangeable.

This commit streamlines the interface of the two mentioned classes so
that their purpose is more clear, and so that their behavior can be more
clearly indicated by their purpose. Methods that did not align with the
ostensible purpose of each class were removed.